refactor(ui): extract the keystroke timeline into a shared panel
Pulls the timeline modal's content — zoom, hover tooltips, legend, and
the line/word rows — into KeystrokeTimelinePanel, which takes an
already-resolved log so a later change can render it inline on the
typing-test completion screen. The stat block unifies both surfaces'
numbers (WPM, KPM, accuracy, KSPC, time, words, overlap, plus the
missed list and error classes) by reusing the completion screen's own
mistake-summary pieces, and the modal shrinks to a shell that fetches
the log.
